+++ FirstCPPApplication/enemies/bad_mother.cpp
@@ -63,7 +63,7 @@ void BadMother::championize()
     // this->representation->setFGColor(TCODColor::green+TCODColor::green+TCODColor::darkYellow, true, false, true);
     this->attrs->health->current_val+=this->attrs->health->current_val;
     this->attrs->health->max_val+=this->attrs->health->max_val;
-    this->xp_value= floor(this->xp_value*1.5);
+    this->xp_value = (int)std::floor(this->xp_value*1.5);
 }
